pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fm-toolkit"
version = "0.1.0"
description = "FX & Rates Pricing Demo toolkit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31",
]

[project.optional-dependencies]
dotenv = [
  "python-dotenv>=1.0",
]
app = [
  "pandas>=2.0",
  "streamlit>=1.32",
]
dev = [
  "pandas>=2.0",
  "pytest>=8.0",
  "streamlit>=1.32",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
